Fix inconsistency in ColorInputPreprocessor
ColorInputPreprocessor is responsible for converting ARGBType images into
three channel FloatType images. It did divide image intensities by a value
of 255 when doing this conversion. This is a bit of an unexpected behavior.
Especially because ImageJ2 also automatically converts ARGBType to
RealType without scaling the image intensities. This led to inconsistent
behavior when the pixel classification was wrapped as a ImageJ2 command.

This commit fixes this problem by removing the image intensity scaling
from the ColorInputPreprocessor. And introduces a new ChannelSetting for
this default behavior.

A DeprecatedColorInputPreprocessor the always scales image intensities
by 1/255 is added to provide backward compatibility. This old behavior
if expected by saved models trained on RGB images.
ChannelSetting.DEPRECATED_RGB supports this behavoir.