Speed up RandomForestPrediction
This introduces class RFPrediction, which flattens the FastRandomForest
into arrays differently than the GPU version.

The random forest is encoded in short[] attributes, float[] thresholds,
and float[] probabilities arrays.

Trees are "flattened out" to full binary trees of their respective height.
For a tree of height h, space for 2^h-1 internal nodes is allocated in
attributes[] and thresholds[], and space for 2^h leaf nodes is allocated in
probabilities[].

For an internal tree node i, attributes[i] and thresholds[i] are the index
of the attribute to select and the threshold to compare it with. If the
attribute is smaller than the the threshold go to the left child, otherwise
go to the right child. The left child of node i is at 2*i+1. The right child
of node i is at 2*i+2.

At the leafs at maximum depth of the tree, no attributes[] and thresholds[]
space is required, instead probabilities[] contains the leaf probabilities.
A value attribute[i] < 0 signifies that this branch of the tree terminates
at node i (early, i is not at full depth). In this case, the leaf
probabilities are stored at the position we would reach by always going to
the left child until we reach full depth.

Trees are ordered by height. So first all trees of height 1 are put into the
arrays, then height 2, etc.

For the first few heights there are special case implementations for
evaluating the tree. For example, for a tree of height 1 we know that it
always has exactly one attribute comparison that directly results in either
of two leaf probability slots.

Also for accumulating leaf probabilities there are special case
implementations for 2 classes and 3 classes.

See RFPrediction.generic_distributionForInstance() for the "simple" version
without special cases.