Convert SegmentationBenchmark to demonstrate RandomAccess slowdown
The benchmark demonstrates, that ArrayRandomAccess is 4x slower
after a image is segmented with imglib2-trainable-segmentation.

Why is that?
I guesse, it's again the problem with the generic function calls.
This time the problem is in the method ArrayRandomAccess.fwd(d).
The code for the feature computation heavily uses this method,
so it gets optmized be the JIT-compiler.
And it's used with different pixel types: FloatType, DoubleType,
UnsignedIntType. Internally the fwd() method calls type.incIndex( ... ).
And this function call gets realy slow.