Segmenter: pre-initialize CPU & GPU optimized random forests
When segmenting a block, there was always an instance of
GpuRandomForestPrediction or CpuRandomForestPrediction created.
This wasted a lot of resources especially with very large random
forest. This change should largely increase the performance of Labkit
when used with large random forest. Under normal conditions the
performance is only slightly better.